requirements.txt
streamlit==1.38.0
langchain==0.3.1
langchain-community==0.3.1
langchain-chroma==0.1.4
langchain-ollama==0.2.2
beautifulsoup4==4.12.3
requests==2.32.3
pypdf==5.1.0
python-dotenv==1.0.1
chromadb==0.5.11
langchain-groq==0.2.0
langchain-huggingface==0.1.0
sentence-transformers==3.1.1
torch==2.4.1